# docs/swagger.yaml
definitions:
...
  requests.KubeApplyYamlRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      yaml:
        type: string
    type: object
...
  requests.KubeDaemonSetCreateRequest:
    type: object
  requests.KubeDaemonSetRollbackRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      revision_name:
        description: |-
          指定回滚到的历史版本（ControllerRevision）
          建议用 revision_name（更直观），如果你喜欢数字 revision 也可以再加一个 int64 字段做二选一
        type: string
    type: object
  requests.KubeDaemonSetUpdateImageRequest:
    properties:
      container:
        description: 目标容器名称
        type: string
      image:
        description: 新镜像地址，例如 nginx:1.27
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
...
  requests.KubeEventListRequest:
    properties:
      continue:
        type: string
      kind:
        description: Deployment/StatefulSet/DaemonSet/Pod/Node...
        type: string
      limit:
        description: 默认 50
        type: integer
      name:
        type: string
      namespace:
        description: 为空=全局
        type: string
      reason:
        type: string
      since_seconds:
        type: integer
      type:
        description: Normal | Warning
        type: string
    type: object
...
  requests.YamlCreateRequest:
    properties:
      yaml:
        description: YAML 内容
        type: string
    required:
    - yaml
    type: object
  response.Response:
    type: object
...
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/daemonset/apply_yaml:
    put:
      consumes:
      - application/json
      description: 应用修改后的 YAML 配置到 DaemonSet
      parameters:
      - description: YAML内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeApplyYamlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 应用 DaemonSet YAML 配置
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeDaemonSetCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 创建 DaemonSet（可选创建 Service）
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/create-from-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: YAML 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.YamlCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 从 YAML 创建 DaemonSet
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/delete:
    delete:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 删除 DaemonSet
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/delete_service:
    delete:
      description: 删除指定命名空间下，与 DaemonSet 同名的 Service 资源
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Service 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除 DaemonSet 对应的 Service
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/detail:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 DaemonSet 详情
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/ds_pods:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 DaemonSet 关联的 Pod 列表
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/events:
    post:
      consumes:
      - application/json
      parameters:
      - description: 事件查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeEventListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 DaemonSet 相关事件
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/history:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 DaemonSet 历史版本
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/list:
    get:
      description: 支持分页、命名空间过滤、名称模糊查询
      parameters:
      - description: 命名空间
        in: query
        maxLength: 100
        name: namespace
        type: string
      - description: DaemonSet 名称(模糊匹配)
        in: query
        maxLength: 100
        name: name
        type: string
      - description: 页码 (从1开始)
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量 (默认20)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 K8s DaemonSet 列表
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/restart:
    post:
      consumes:
      - application/json
      description: 触发指定命名空间下 DaemonSet 的滚动重启（等价于 kubectl rollout restart ds <name>）
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DaemonSet 重启成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 重启 DaemonSet
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/rollback:
    post:
      consumes:
      - application/json
      description: 将 DaemonSet 回滚到指定的历史版本（ControllerRevision）。不传可在服务端实现“回滚到上一个版本”的兜底策略（可选）。
      parameters:
      - description: 回滚参数（namespace、name、revision_name）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.KubeDaemonSetRollbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: DaemonSet 回滚成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 回滚 DaemonSet
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/update_image:
    put:
      consumes:
      - application/json
      description: 修改指定命名空间下 DaemonSet 的容器镜像（支持滚动更新）
      parameters:
      - description: 更新镜像参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeDaemonSetUpdateImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新 DaemonSet 容器镜像
      tags:
      - K8s DaemonSet 管理
  /api/v1/k8s/daemonset/yaml:
    get:
      description: 获取指定 DaemonSet 的 YAML 配置
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: DaemonSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 DaemonSet YAML 配置
      tags:
      - K8s DaemonSet 管理
...
swagger: "2.0"